+++ EventsController/ChangeLog.md
@@ -1,2 +1,4 @@
-﻿## Version 1.0.0
-- Initial release
+﻿## Version 1.1.0
+- All events are now on by default.
+- Added a new features in the seasonal events tab called Weather based fire the parameters of fires will increase and decrease based on the weather and weather history 
+- fixed some settings not working correctly